[CLEANUP] Move shared controller tests into an abstract base class
This get rids of loads of duplicated code.

Fixes #309
